Test and handle bogus patient_id
Testing for non-existent patients and bogus patient_ids crashed.
The "do { } while(i < list.length)" construct fails on empty lists.
This converts "do {} while()" into "for (i = 0; i < list.len; ++i)"
which handles empty lists.

If it is an error that a list is ever empty, lets write an assertion
for that.