Move "window.addEventListener('load',...)" to page
Rather than have the inclusion of the "common.js" script do the
magic of adding the on-load listener, it now defines a function
to be included by the page.

Each page which includes "common.js" now also has an explicit
"window.addEventListener('load',..." call which includes the
new function.

With this, we are now able to have more control of the order of
page setup and websocket setup.

Additionally, "common.js" should now be a little easier to test.