prep: add medication-advice-list to five-pages.js
This commit creates a 'thirdparty' directory with links to the
(minimized) 'showdown' library so that it can be served to the
browser and used client-side.

The tar-file creation (with the transform) does not work as one
might hope, thus a work-around (re-linking on the install machine)
can be seen in vm-init.

The giant medication-advice-list code can and should be split into
multiple functions both for readability and also re-use in the
consult view.