updated TODO
Cannot build tests for the previous commit (copying EHR and Patient text to clipboard) because TestCafe cannot check the contents of the clipboard. A workaround is to paste the result in a free text box, but our current free text boxes do not support multiline texts. This is probably not ideal - it would be better to support multiline texts in the two "other" text boxes. Added this to the TODO.